CORE-9570 Refactor the group parameters schemas to support persisting MGM signature information on a Member's cluster (#896)
This change is necessary so that members can persist and keep a record of the MGM's signature over a set of group parameters. Member's will only be able to access this signature through internal Corda code.

The GroupParameters.avsc schema has been replaced with SignedGroupParameters, and PersistentGroupParameters has been added as the schema for persisting group parameters to a compacted kafka topic as viewed by a specific identity. This is similar to the schema structure for member info. SignedGroupParameters contains a serialised KeyValuePairList for the group parameters, and optionally the MGM's signature. The signature should only be null if the owner of that data is an MGM. The group parameters must be held in their serialised form for signature verification.

The group parameters DB table has been extended to include optional signature information. Again, this should only be null for the MGM.

This PR also includes a minor tidy up to remove WireCpiAllowList, WireCpiVersion, and CpiVersionEntries since they are unused.

Matching runtime-os changes can be found here: corda/corda-runtime-os#3267