CORE-9570 Refactor group parameters to include MGM signature information and a hashing function (#3267)
This PR persists and exposes the MGM's signature over group parameters for members in dynamic and static networks. The MGM still persists and handles group parameters as unsigned and signs them on distribution. There will be a follow up ticket to change that so that the MGM also persists the group params as signed. At that point, group parameters can be handled everywhere as signed, but until then, the membership group reader returns both signed and unsigned. Only MGM code should use unsigned, and everywhere else should use signed.

In order to support signing of static network parameters, a fake MGM is now created and pushed to the group policy file during CPI upload. The signing key information is persisted to the database for signing operations. This is only possible for static networks because these are not production network and are purely for test purposes. The signing key information being pushed to the database makes it even more important now that static networks do not go to production in their current form. Static network information was previously pushed to the message bus, but has now been move to the database to allow for transactional writes along with CPI upload when a CPI is for a static network.

Supporting API changes can be found here: corda/corda-api#896

This also adds a notary member to the e2e smoketests to force group parameters updates other than the initial distribution during the smoke tests. Uncovered a bug in the multi cluster environment blocking this change so I raised CORE-11878 to address that issue.