CORE-7994 Replace key with key id in digital signature packet (#956)
At the moment, the `SigningService.sign` returned signature packet includes the signing key (public key) and the signature packet will be sent over the wire to counter parties. This means that currently, the receiver of that signature packet can grab that key and use it directly to verify the signature. However, they should be checking that this key exists in the locally known/ trusted counter parties keys. By changing the key to key id it means the consumer of the signature has to to resolve the key out of the key id to proceed with signature verification.

* Replaces `DigitalSignature.WithKey` with `DigitalSignature.WithKeyId`

* Adds signing key to `TransactionSignatureService.verifySignature` parameters, since the mapping from key to key id should take place in `TransactionSignatureService.verifySignature` caller sites

---------

Co-authored-by: Kyriakos Tharrouniatis <kyriakos.tharrouniatis@r3.com>