CORE-7994 Replace key with key id in digital signature packet (#3400)
At the moment, the `SigningService.sign` returned signature packet includes the signing key (public key) and the signature packet will be sent over the wire to counter parties. This means that currently, the receiver of that signature packet can grab that key and use it directly to verify the signature. However, they should be checking that this key exists in the locally known/ trusted counter parties keys. By changing the key to key id it means the consumer of the signature has to to resolve the key out of the key id to proceed with signature verification.

* Refactors `TransactionSignatureService.verifySignature` and use it in `UtxoSignedTransactionImpl`, `ConsensualSignedTransactionImpl` and `UtxoFinalityBaseV1` so as to go through the transactions for signature verification because they hold the signatories and therefore that is where we can map the signature's key id to keys

* Maps key id to key in `DigitalSignature.WithKeyId` consuming sites

* Statically converts key to key id in `SigningServiceImpl.sign` return type (`DigitalSignature.WithKeyId`)

* Internalizes `DigitalSignature.WithKey` to continue using it internally