CORE-8712 - Switch to use MemberInfo with latest serial when processing initiator hello (#3469)
Switch to use MemberInfo with latest serial when processing initiator hello.

At the point we receive the session initiator hello, we don't know the identity on our side the session is established with. This is only done at the later phases of the session handshake. Previously, we were picking the Member Info as viewed from a randomly selected locally hosted identity (the first one) up until that point. The new logic here picks the Member Info with the latest serial, so that we make sure we use the latest view. This will be a bit more tolerant against cases where peers have upgraded their Member Info and some locally hosted identities are on stale versions.