CORE-1068 Introduce Packaging Test Application (#3428)
This is a simple ledger test application which will form the basis of packaging testing for multiple versions of contracts. Its intended use is something along the lines of:

Upload v1 of the app to Cluster B
Create vnode Bob in Cluster B
Upload v2 of the app to Cluster A
Create vnode Alice in Cluster A
v1 of the app will mint states for Bob on Cluster B, verifying v1 of the contract works for Bob
Transfer states to Alice on Cluster A, who has v2 contract, which should fail
Upload v1 of the app to Cluster A
Transfer states to Alice on Cluster A, who now has access to v1 contract, so now should pass
State transfers for each vnode can be tested with the "report states" flow
At the moment this is a single Cordapp. A clone must be made and built to create v2 of the application.

To mint states you execute the flow like so, where the list of integers is a list of state values to mint:

{
    "startFlow": {
        "clientRequestId" : "q3",
        "flowClassName" : "net.cordapp.testing.packagingverification.MintFlow",
        "requestBody" : "{ \"stateValues\" : [10, 20] }"
    }
}

To transfer states execute the following flow, specifying the issuer, recipient and the amount to transfer. Note there is little error checking so don't try to trick the application:

{
    "startFlow": {
        "clientRequestId" : "p1",
        "flowClassName" : "net.cordapp.testing.packagingverification.TransferStatesFlow",
        "requestBody" : "{ \"issuerX500Name\" : \"C=GB, L=London, O=Bob\", \"recipientX500Name\" : \"C=GB, L=London, O=Alice\", \"value\" : 10 }"
    }
}

To report states in any vnode, execute the following flow:

{
    "startFlow": {
        "clientRequestId" : "t5",
        "flowClassName" : "net.cordapp.testing.packagingverification.ReportStatesFlow",
        "requestBody" : "{}"
    }
}