CORE-11878 Add additional members to test to prevent regression when onboarding multiple members across clusters (#3500)
CORE-11878 has already been resolved by CORE-11922. No additional fix is required. This PR updates an e2e test to add additional members in the multi cluster scenario to attempt to prevent regression in this area again since updating this test was blocked originally by this bug.

Note: this test should be ported over to the e2e test repo but that will be covered as part of CORE-10087 and is not in scope for this PR.