CORE-12083 - Updated the avro schema `VirtualNodeInfo` and the database schema `config-creation-v1.0.xml` (#1002)
CORE-12083 - Added the externalMessagingRouteConfig field to the avro schema VirtualNodeInfo.
Added the column externalMessagingRouteConfi column to the virtual_node.
The externalMessagingRouteConfig will contain the route configuration for external messaging. The field is a nullable string which will contain a JSON structure. The field can be null meaning no configuration has been provided.