CORE-11502: Behaviour change: use the tenant ID specified by the context parameter when creating a wrapping key, rather than always storing wrapping keys in the CRYPTO tenant in the cluster datbaase.
Split out wrapping key material into its own table so that we can store multiple representations at once, for later use during key rotation when that is implemented. (Allowing easy joins motivated the behaviour change on where the wrapping keys are stored, but the behaviour change itself is appropriate.)

Add a rotation date column in the wrapping keys, initially set to a long time in the future. Nothing yet checks the rotation date.

Allow a set of root wrapping keys to be passed into the crypto processor, with one selected to be the default. This required supporting secrets in arrays during validation.

Now that we have removed all OSGi and Lifecycle logic from the repository layer of the crypto processor it it possible to have fast integration tests that consturct an in-memory database directly and exercise the repository layer code, and can be debugged without needing to go multi-process. So, rework the RepositoryTests like that. Take the opportunity to group them according to the repository class they are testing, in the standard structure.

Stop allowing querying the signing repository by master key alias (aka the warpping key), since it is possible with the new strucutre that there are multiple master keys for the same signing key.

Use data classes for the *Info DTOs, and take care over their equqals and hashCode methods where ByteArray is concerned.

Make the repository operations return the DTO, to align better with other JPA code in Corda. This might sometimes make calling code more compact since it can use the Fluent style.

Disable certain PersistenceTests for now, which are mostly duplicates of the fast new non-OSGi integration tests.

Disable unit tests for the SigningRepository, since their mocking requires updates.

Add some tracing code (at trace log level) for wrapping key handling.

Fix typo in TestWrappingRepoitory filename.