CORE-12190 - Increase gateway default timeout for requests (#1030)
The current default timeout for HTTP requests from the p2p gateway was 1 second, which is a bit tight for some scenarios. For example, our test infrastructure is not very well resourced and there were a lot of timeout exceptions in the logs. Raising the defaults slightly to make this a bit more tolerant to delays.

Conscious we also have e2e replay configuration at the link manager, but I intentionally didn't go down the path of adjusting this as that might require some fine tuning to be done properly and the intention of this change is to mainly reduce noise in our logs for easier troubleshooting.