CORE-12233 - Improve error messages for invalid endpoint (#3546)
This change improves the error messages shown when one of the endpoint URLs are invalid. Previously, the order of checks was slightly incorrect and if someone provided a URI with an invalid host, they would get an error about port (see related ticket for more details). This re-orders the checks and adds an extra check for a completely invalid URL.