CORE-11948: State synchronization enhancements in the state and event message pattern (#3542)
This PR introduces two enhancements to try and address slow state and event rebalances. Firstly, the StateAndEventConsumer now keeps track of all partitions currently being synchronized, and all partitions that are currently in sync. By doing this, the event consumer poll can be avoided if it has been invoked recently and there are no currently in-sync partitions. Secondly, the state consumer is now unassigned partitions when they are in sync, which allows polls of the state consumer to only return data needed to update the in-memory state.