CORE-11836 - Hide mgm plugin commands that are not ready for external use yet (#3570)
Hiding mgm plugin commands that are not ready for external use yet from the usage help of the parent command. Note that these commands will still be executable, so that we can use them for internal testing but they shouldn't be discoverable by end users that don't know about them and so they will be considered unsupported at the moment. Also, I elected to not hide subcommands of `onboard` as well, so that it's a bit easier to use internally by us.