CORE-9790: refactor helm bootstrap to reduce repetition (#3597)
Previously we had 6 instances of a pattern where we run corda-cli to make some SQL, then run another container to execute it. This was hard to maintain, so used a new named template to go from 6 to 1 copy of the declarations.

To verify this produces the exact same output, I used helm template and then diffed the output against the old run. The differences are only in the random elements.

This exercise has revealed various unnecessary consistencies in the whitespace, quoting and naming. For now, I've got conditionals that preserve the inconsistencies, and we should make this big change, then remove the inconsistencies in later work.