Change MemberSynchronisationServiceImpl + MgmSynchronisationServiceImpl
To work in the same way as the DistributeMemberInfoActionHandler.
On the Memberside we only hash our own member info, if we have
been suspended. On the MGM side only distribute a member list with
our own Member info. Also delete GroupParametersUtils.kt (GroupParametersNotaryUpdater
is used instead. Improve some of the unit tests.