CORE-12498 Vault named query parsing escape special characters
`?` and `::` need to be used in Postgres JSON queries; however, they
cause parameter issues when creating the native query and execution
errors when executing them.

To bypass these errors these characters must be escaped in the executed
query.

We have decided to keep our queries tidy so the escaping has been hidden
within our parsing code. This
means that developers write `?` and `::` into their query
strings, and we translate them into `\\?\\?` and `::` respectively.