CORE-8962 flow mapper absorbs messages when in state closing (#3576)
When a session has completed the Flow Mapper holds onto its state and sets it to CLOSING. It now ignores any incoming messages. This means if an ACK message is lost that one party cannot complete its close and will log the session terminated in error while the other party will record it completed successfully. This change will send automatic ACKS in that situation to prevent this. This PR also introduces a new state, ERROR.