CORE-10576 Trigger member list distribution when suspending and activating a member (#3565)
Trigger member list distribution from the rest worker after updating the database when suspending or activating a member. Additionally change the db transactions to remove and add a notary from the group parameters if a member is suspended/activated. Also changed the distribution protocol inside DistributeMemberInfoAction to only send their own member info to a suspended member (so information is not leaked about the state of the network). Also changed the Member and MGM side of the Synchronisation service to reflect this.