CORE-10419 Passing Facade ID and Facade Method in SessionInit (#3603)
P2P SessionProtocol uses SessionInit message to indicate which flow class to run (a class of the responder flow), while SessionData message carry on ... "data"/payload. An initial thought was we will have a single message with a FacadeId/FacadeMethod and the payload, however reusing P2PSession protocol with minimal changes to Handler, we quickest solution is to split a request data into:
SessionInit - carries FacadeName and FacadeMethod in the session properties which is a map/list inside SessionInit
SessionData - carries a payload.
For E2E testing IntertopProcessor has been modified to read an Inbound SessionInit and use InteropProcessor's inner mapper to get the FlowClass name, and then the class name is added to SessionInit message, then FlowProcessor starts an interop flow skipping the "protocol" value and using "flowClass" from SessionInit. This however still uses CPI hash as normal P2P flows, so initiating and initiator flows needs to be on the same Cordapp still. This proves that the data set in an initiating flow is interpreted and used to run the responder flow.