CORE-10585 Add E2E test utilities to support member suspension testing (#3581)
This adds the e2e test utility functions used in the e2e test repo for testing membership suspension and activation.

This includes a refactor to make utility functions operate in the context of ClusterInfo rather than passing the ClusterInfo as a parameters. This allows for more readable tests, especially for multi cluster tests.