CORE-9790: helm corda-lib support for Vault, and tidy up (#3610)
Support Hashicorp Vault for config secrets if it is configured, while preserving the Corda-native configruation system for non-Vault users. Tested for both Corda Open Source, and with corda/corda5-enterprise#38 for Corda Enterprise. The enterprise/Vault scenario requires careful handling of secrets.

Add a sequence number prefix to each init container, so that they are listed in the right order and the ordering requirement is clear, since this initially confused me. This seems like a nice usability benefit for operators.

Add a short script to deploy to kubernetes in one go, respecting the prereqs chart and the requirement to delete namespaces.

In my earlier change I was careful to preserve all the idiosyncancies of the Helm output, so I knew that fiddly refactoring would have no impact on the Kubernetes resources that helm specifies. Now that's merged, we no longer need to preserve the exact whitespace and inconsistent use of quoting, so that allows the Helm chart to be simplified.