CORE-11959: make lifecycle being down cause retries in crypto flow ops
To that end, make the error thrown by AbstractComponent when invoking operations on non-UP components be a specific type, since it was IllegalStateException and we don't know if it is appropriate to retry this.

I believe there are situations where errors may not be recoverable, and we will rely on the operator, often via Kubernets, to restart the worker, e.g. because the health check is returning negative.