CORE-11959: make lifecycle being down cause retries in crypto flow ops (#3712)
To that end, make the error thrown by AbstractComponent when invoking operations on components in state DOWN or ERROR be a specific type, since it was IllegalStateException and we don't know if it is appropriate to retry this.

I believe there are situations where errors may not be recoverable, and we will rely on the operator, often via Kubernets, to restart the worker, e.g. because the health check is returning negative. However, if we error out from the CryptoOpsClient into CryptoFlowOpsBusProcessor, then a platform error will be reported to the flow. If we wait, maybe we can proceed either after the service is restarted (e.g. by Kubernetes) or if the lifecycle goes UP again some other way. For instance, if we get a UnknownHostException from Kafka and put CryptoOpsClient down, then without this change we will error out a bunch of flows. With this change, we will wait until the service is restarted and then maybe succeed (if the TTL hasn't expired), or potentially we can have some additional retry logic that marks the CryptoOpsClient UP again when Kafka comes back.

Long term, we may remove the AbstractComponent abstraction, and that takes work in each component that uses it, so this is a short term small fix to improve robustness in 5.0.