CORE-12364 cleanup nonexistent state (#3689)
When a flow is started by the REST api a start flow event is generated by the FlowRestResourceImpl and it is passed to the Flow Mapper with the key set to the FlowKey (converted to string). This allows the flow mapper to map the client requested flow to a FlowId. When a flow fails or finishes the flow engine will schedule a cleanup event for this key on the mapper.

For flows which are triggered by session initiation (from receiving a SessionInit) then this key will not exist. Therefore we should not send this cleanup event in FlowFinishedRequestHandler and FlowFailedRequestHandler when the flow was started in this way.