CORE-13030: Add a batch publish API (#3720)
The flow REST resource can show some thread contention when lots of clients are trying to publish start flow events simultaneously. This occurs as the publisher in the REST resource is transactional to ensure that the start flow and status events are published together. However, by enabling transactionality, the publisher only allows a single thread to publish its records at once, resulting in a bottleneck.

To address this, a new API is added to the publisher that allows requests from multiple threads to be grouped together in a single batch when transactionality is enabled. This works by adding the requests to a queue, and then waiting to acquire a lock. When the lock is acquired, the full queue is drained and all records are published in a single transaction. This means a thread may publish records from multiple threads at once.

This change shows an improved start flow latency when running many clients in the soak tests. Previously the median time was 800ms to respond to a start flow request with 96 clients. With this change, response times are reduced to an order of 80ms.

Also, adjusted the state and event processing time metric to record the time to process a (sub-)batch rather than an individual event. This is much easier to interpret when looking at metrics coming out the cluster.