CORE-12552 Follow up from CORE-10576 (#3752)
Functional changes:
Change the Member Synchronisation Service to do a look up in known members if the member is not found in the update package. Use the status in the same way, as if it was in the membership package to calculate the expected hash.
Add logging to ActivateMemberHandler and SuspendMemberHandler.
Use a stream instead of a list in the large queries (for notary updates the group parameters).

Non Functional changes:
Improve unit test coverage of the MGM Synchronisation service. Add coverage of notary suspension/activation to SuspendMemberHandlerTest and ActivateMemberHandlerTest.
Test notary activation in the MembershipPersistenceTest.
Unit test changes from CORE-10576 in the MgmSynchronisation service.
Move coverage of exceptions from the AddNotaryToGroupParameterHandlerTest to the GroupParametersNotaryUpdaterTest. To make purer unit tests.