CORE-12911 Use only class bundle name for evolved classes (#3780)
Ultimately Corda 5 should use Corda-CPK-Cordapp-Name from the CPK metadata to identify a class type as being the same as another version of that class. See CORE-13259. This will be written in the evolvable tag. This change goes part way to supporting that in the evolvable tag.

Right now CorDapp developers cannot set this metadata, the gradle plugin always sets it to bundle symbolic name, so offering full support for Corda-CPK-Cordapp-Name is not yet possible, but this change sets up EvolvableTag and everything around it to support that in later versions.