CORE-8714 Set the Protocol and PKI mode after the InitiatorHandshake instead of the InitiatorHello (#3834)
We only know the initiator's identity after we have received the InitiatorHandshake message. The previous behaviour was looking for any identity in the same group and then using that to look up the group policy. This will cause problems once updating the parts of the group policy used by the Link Manager is supported. e.g. reregistering and updating the protocol mode.

To do this I added a new step into the responders authentication protocol VALIDATED_ENCRYPTED_EXTENSIONS. This step comes after validating the InititaorHandshake message. I moved selecting and checking for a valid protocol mode and validating the certificate from validating the initiator handshake message, to this new stage.

I left in the group policy lookup when processing the InitiatorHello message, this is only used to lookup the network type (Corda 5 or Corda 4). We could consider hard coding it to Corda 5.