CORE-12581: run signing service direct from flow operations processor rather than going via RPC, in order to eliminate problems with lack of robustness in crypto to crypto RPC calls (#3801)
Perform flow crypto operations direct from the thread that picks up the Kafka topic, rather than crossing Kafka twice more.

Internally, the crypto worker implements a signing service, which provides operation to sign and verify signatures, as well as other operations such as creating and retrieve asymmetric keys. This exposed to Corda via 2 bus processor, each watching a Kakfa topic. One topic and bus processor is for flow operations, and the other for RPC operations. Up to this point, the bus processor for flow operations would handle requests by posting them by being RPC client, so posting to the RPC topic and waiting for a partition owned by that instance of the crypto processor. The trouble is that this is unreliable, since Kafka topics can and are rebalanced at any point, and there was no retry code.

This pattern of messages crossing kafka twice in each direction was done to allow, in the future, different instances of the crypto processor to handle a subset of keys. By posting crypto flow operations back to Kafka, any arrangement we have to make sure that a crypto processor instance with access to a specific key would extend to flow operations as well. However, we now think that for the foreseeable future all crypto processor instances will be able to access any key, and therefore this sophistication is not necessary, or, due to its inherent complexity and performance overheads, desirable.

Previously, we had 2 separate OSGi+Lifecycle service classes that subscribe to Kafka. Since they both now need to access the signing service, and since the signing service is not itself OSGi, that means that we would have to construct the signing service implementation class. It's also a lot of code, and in particular code using AbstractConfigurableComponent. Instead, we have Crypto Processor itself receive dynamic config then construct the Signing Service and the Kafka subscriptions. This is much simpler, plus now the crypto processor itself actually does something more than report the overall health of the crypto processor as ready once the dependencies start. Rejected design alternatives were to create the signing service twice, each with its own cache, or make the signing service be OSGi+Lifecycle itself, which makes it harder to test.

Make the crypto service bus implementations exported, so that crypto processor can import them. Perhaps we could alter move the service bus implementations into crypto processor, but that can be done in a separate PR to keep this reviewable and minimise changes if this is backported.

I notice that FilterMyKeysFlowQuery avro message is unused, so have added a separate JIRA CORE-13554 to address this.

Rework the crypto flows operations bus processor unit tests. There was a lot of repetition of the machinery to actually run the flows, mixed up with the assertions for specific tests which made it pretty hard to maintain, and the code all had to be touched since the underyling code is the signing service now rather than the crypto ops client.