CORE-12555 Interop Facades as Corda service (#3792)
The aim of this PR is to demonstrate a working Client Proxy WEFT and Server Dispatcher WEFT from within Cordapps and a sandbox. The current Corda API is simplified for now.
No changes to security policy files are needed. Three occurrences of programmatic AccessController.doPrivileged(PrivilegedExceptionAction { ... }) around Weft operation have been added. Especially
wrapping a proxy instance creation as we expect creating dynamic proxies to require privileges as well. Rational SecurityException - if a security manager, s, is present and any of the following conditions is met:
- the given loader is null and the caller's class loader is not null and the invocation of s.checkPermission with RuntimePermission("getClassLoader") permission denies access;
- for each proxy interface, intf, the caller's class loader is not the same as or an ancestor of the class loader for intf and invocation of s.checkPackageAccess() denies access to intf;
- any of the given proxy interfaces is non-public and the caller class is not in the same runtime package as the non-public interface and the invocation of s.checkPermission with ReflectPermission("newProxyInPackage.{package name}") permission denies access.