CORE-5555 - Token Selection - Implement Balance Query API (#3728)
The feature balance query was added in. This new feature allows people to query the balance of a pool of tokens. There are two different values representing the balance. One of the values is the balance, or the available balance, and the second one is the total balance. The logic behind these two values is similar to a bank account which always has the available balance and the total balance including charges that had been ordered but have not been executed yet.

Tests were added to provide some reassurance that the code is working as expected.