CORE-13870: Unique Kafka ClientId
The 'client.id' property is built aggregating multiple ids,
'$subscriptionType-$group-$topic-$idCounter'. All components are static
except for 'idCounter' (instance of 'AtomicLong'), which is supposed to
be unique within a single JVM. When multiple instances of the same
worker are started, however, there's a chance of ending up having the
same 'client.id' for different consumers within the same consumer
group, which makes troubleshooting difficult and can cause issues in
the Kafka rebalance protocol.

- Improve uniqueness guarantee for the 'clientId' by using 'UUID'
  instead of 'AtomicLong'.
- Remove redundant information ('group.id') from logger names, logging
  messages and Kafka 'client.id' configuration.
- Standardise logger name (class name + clientId) across subscriptions.