CORE-13755 Add persistence metrics for membership (#3906)
Added metrics to capture both the overall membership persistence handler execution time and the membership persistence transaction times. This is because the recent performance issues we saw in the e2e test environment was caused by the time spent getting the database connection which only became clear after comparing the transaction times to the handler execution times.