CORE-13987 Ledger metrics (#3939)
Add metrics for:

- `flow.crypto.time`
- `ledger.flow.verification.time`
- `ledger.verification.contract.total.time`
- `ledger..verification.contract.time`
- `ledger.verification.contract.count`
- `ledger.flow.persistence.time`
- `ledger.backchain.resolution.chain.length`
- `serialization.amqp.serialization.time`
- `serialization.amqp.deserialization.time`

This involved wiring in `CurrentSandboxGroupContext` to a lot of services and making it available in all sandboxes.

The `SeriazliationService` implementations in `:flow-service` and `SerializationServiceProxy` in other sandboxes provide sreialization metrics without adding directly to the serialization library.

`recordSuspendable` added to allow recording within fibers, which is needed do to `Timer`s being closable and failing serialization.

`recordOptionally` added to allow recording of `Timer` metrics that are only recorded if they surpass a certain duration. This is used by the serialization metrics.