CORE-13858 - Record inbound message rate in the loopback case (#3962)
Record inbound message rate for the loopback case as well (i.e. when destination is locally hosted within the same cluster). I also adjusted the code to populate all the tags for that metric in all cases, because otherwise the metric is not exported because the prometheus client used by micrometer requires the same set of tags to be populated for a specific metric name when within the same process.