CORE-13952: Remove tracking of RPC subscription (#3924)
* CORE-13952: Remove tracking of RPC subscription

We shouldn't need to track the subscription as part of the lifecycle. This can lead to spurious ERROR messages when we don't correctly manage the tracking of the subscription and our handling of the subscription itself. (e.g. we bring the subscription down but we haven't stopped tracking it)

Removing the subscription tracking clears that up and ensures we're consistent with the rest of Corda as well.